requirements.txt
# Turkic Language Processing Requirements
# Install with: pip install -r requirements.txt

# ============================================================================
# MINIMAL INSTALLATION (for using pre-trained models)
# ============================================================================
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
openai-whisper
torch>=1.10.0
joblib
mteb
# ============================================================================
# OPTIONAL DEPENDENCIES
# ============================================================================

# mteb

# For FastText model training
# Uncomment to install:
# fasttext>=0.9.2

# For Transformer models (BERT, XLM-RoBERTa)
# Uncomment to install (requires ~2GB download):
# transformers>=4.20.0
# torch>=1.10.0
# datasets>=2.0.0
# accelerate>=0.20.0
# torchvision
# torchaudio

# For Jupyter notebooks
# Uncomment to install:
# jupyter>=1.0.0
# notebook>=6.4.0

# For visualization (optional)
# matplotlib>=3.7.0
# seaborn>=0.12.0

# ============================================================================
# INSTALLATION EXAMPLES
# ============================================================================

# Basic usage (pre-trained model):
#   pip install numpy pandas scikit-learn

# FastText training:
#   pip install numpy pandas scikit-learn fasttext

# Transformer training:
#   pip install numpy pandas scikit-learn transformers torch datasets accelerate

# Full installation:
#   pip install numpy pandas scikit-learn fasttext transformers torch datasets accelerate jupyter

# ============================================================================
# NOTES
# ============================================================================
# - Start with minimal installation to use the pre-trained model
# - Add FastText only if you want to train FastText models
# - Add Transformers only if you want to train BERT-like models (GPU recommended)
# - torch is large (~2GB) - only install if training Transformers
